Updated Bookmarks and Twitter icon

+++ frontend/src/MainPage/Sidebar/Bookmarks.jsx
@@ -6,11 +6,12 @@ import {
   Typography,
   Menu,
   MenuItem,
+  Snackbar,
 } from "@mui/material";
 import MoreHorizIcon from "@mui/icons-material/MoreHoriz";
+import CheckCircleOutlineIcon from "@mui/icons-material/CheckCircleOutline";
 import { AuthContext } from "../../AuthenticationSystem/AuthenticationSystem";
 import axios from "axios";
-import { useNavigate } from "react-router-dom";
 import BookmarksTweets from "./BookmarksTweets";
 axios.defaults.headers.common["Authorization"] = `Bearer ${localStorage.getItem(
   "token"
@@ -19,12 +20,27 @@ axios.defaults.headers.common["Authorization"] = `Bearer ${localStorage.getItem(
 export const Bookmarks = () => {
   const [posts, setPosts] = React.useState([]);
   const { user } = React.useContext(AuthContext);
-
-  const navigate = useNavigate();
   const [anchorEl, setAnchorEl] = React.useState(null);
   const open = Boolean(anchorEl);
   const [anchorElPost, setAnchorElPost] = React.useState(null);
   const openPost = Boolean(anchorElPost);
+
+  console.log(posts.length);
+
+  const [state, setState] = React.useState({
+    openSnackbar: false,
+    vertical: "top",
+    horizontal: "center",
+  });
+  const { vertical, horizontal, openSnackbar } = state;
+
+  const handleClickSnackbar = (newState) => {
+    setState({ ...newState, openSnackbar: true });
+  };
+
+  const handleCloseSnackbar = (event, reason) => {
+    setState({ ...state, openSnackbar: false });
+  };
 
   const handleClick = (event) => {
     setAnchorEl(event.currentTarget);
@@ -181,10 +197,15 @@ export const Bookmarks = () => {
 
   const handleClearAllBookmarks = () => {
     try {
+      if (posts.length === 0) {
+        setAnchorEl(null);
+        return;
+      }
       axios.delete("http://localhost:3000/tweet/clearallbookmarks");
       setPosts((prevPosts) =>
         prevPosts.filter((post) => post.bookmarked === false)
       );
+      setAnchorEl(null);
     } catch (error) {
       console.error("An error occurred:", error);
     }
@@ -216,50 +237,92 @@ export const Bookmarks = () => {
               @{user?.username}
             </Typography>
           </Stack>
-          <SvgIcon
-            sx={{
-              cursor: "pointer",
-              fontSize: "20px",
-              mr: 1,
-            }}
-            onClick={(event) => {
-              handleClick(event);
-            }}
-          >
-            <MoreHorizIcon />
-          </SvgIcon>
+          {posts.length !== 0 ? (
+            <SvgIcon
+              sx={{
+                cursor: "pointer",
+                fontSize: "20px",
+                mr: 1,
+              }}
+              onClick={(event) => {
+                handleClick(event);
+              }}
+            >
+              <MoreHorizIcon />
+            </SvgIcon>
+          ) : null}
         </Stack>
       </Grid>
-      <Stack>
-        {posts.map((post) => (
-          <BookmarksTweets
-            key={post.id}
-            firstName={post.firstname}
-            lastName={post.lastname}
-            username={post.username}
-            is_verified={post.is_verified}
-            creation_date={post.creation_date}
-            content={post.content}
-            profile_picture={post.profile_picture}
-            likes={post.likes}
-            retweets={post.retweets}
-            image_url={post.image_url}
-            id={post.id}
-            isLiked={post.liked}
-            isRetweeted={post.retweeted}
-            handleDeletePost={handleDeletePost}
-            handleLikePost={handleLikes}
-            handleUnlikePost={handleUnlike}
-            handleRetweet={handleRetweet}
-            handleRemoveRetweet={handleRemoveRetweet}
-          />
-        ))}
-      </Stack>
+      {posts.length !== 0 ? (
+        <Stack>
+          {posts.map((post) => (
+            <BookmarksTweets
+              key={post.id}
+              firstName={post.firstname}
+              lastName={post.lastname}
+              username={post.username}
+              is_verified={post.is_verified}
+              creation_date={post.creation_date}
+              content={post.content}
+              profile_picture={post.profile_picture}
+              likes={post.likes}
+              retweets={post.retweets}
+              image_url={post.image_url}
+              id={post.id}
+              isLiked={post.liked}
+              isRetweeted={post.retweeted}
+              handleDeletePost={handleDeletePost}
+              handleLikePost={handleLikes}
+              handleUnlikePost={handleUnlike}
+              handleRetweet={handleRetweet}
+              handleRemoveRetweet={handleRemoveRetweet}
+            />
+          ))}
+        </Stack>
+      ) : (
+        <Stack p="28px 0px 0px 128px" width="35vw">
+          <Typography variant="h5" fontWeight="bold" fontSize={32}>
+            Save posts for later
+          </Typography>
+          <Typography variant="span" color="rgb(83, 100, 113)">
+            Bookmark posts to easily find them again in the future.
+          </Typography>
+        </Stack>
+      )}
+
       <Menu anchorEl={anchorEl} open={open} onClose={handleClose}>
-        <MenuItem onClick={handleClearAllBookmarks}>
+        <MenuItem
+          onClick={() => {
+            handleClearAllBookmarks();
+            handleClickSnackbar({ vertical: "top", horizontal: "center" });
+          }}
+        >
           Clear all Bookmarks
         </MenuItem>
       </Menu>
+      <Snackbar
+        open={openSnackbar}
+        anchorOrigin={{ vertical, horizontal }}
+        message={
+          <Stack direction="row" spacing={1} alignItems="center">
+            <SvgIcon>
+              <CheckCircleOutlineIcon />
+            </SvgIcon>
+            <Typography variant="span" sx={{ fontWeight: "bold" }}>
+              Cleared all bookmarks !
+            </Typography>
+          </Stack>
+        }
+        autoHideDuration={2000}
+        onClose={handleCloseSnackbar}
+        key={vertical + horizontal}
+        sx={{
+          "& .MuiSnackbarContent-root": {
+            backgroundColor: "#1DA1F2",
+            color: "#000",
+          },
+        }}
+      />
       <Menu anchorEl={anchorElPost} open={openPost} onClose={handleClosePost}>
         {user.username !== posts[0]?.username ? (
           <Stack>
